[Embedded] Harness Router, Inline Space Creation, Bootstrap, Harness/Admin User Setup (#28)
Adds the basic for harness embedded mode:
- Harness dedicated router with custom APIHandler
- Inline Space Creation
- Client for Account/Org/Project
- Bootstrap (Allows for automated creation of admin user and gitness service (used for all platform required ops))
- Inline harness service principal creation
- Ignore flag for ACL.

+++ internal/store/database/migrate/postgres/0001_create_table_paths.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS paths (
 ,path_targetType  TEXT
 ,path_targetId    INTEGER
 ,path_createdBy   INTEGER
-,path_created     INTEGER
-,path_updated     INTEGER
+,path_created     BIGINT
+,path_updated     BIGINT
 ,UNIQUE(path_value)
 );
