+++ internal/store/database/migrate/sqlite/0001_create_table_principals.up.sql
@@ -6,8 +6,8 @@ principal_id              INTEGER PRIMARY KEY AUTOINCREMENT
 ,principal_admin          BOOLEAN
 ,principal_blocked        BOOLEAN
 ,principal_salt           TEXT
-,principal_created        INTEGER
-,principal_updated        INTEGER
+,principal_created        BIGINT
+,principal_updated        BIGINT
 
 ,principal_user_email     TEXT
 ,principal_user_password  TEXT
