+++ internal/store/database/migrate/sqlite/0001_create_table_tokens.up.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS tokens (
 ,token_type           TEXT COLLATE NOCASE
 ,token_name           TEXT
 ,token_principalId    INTEGER
-,token_expiresAt      INTEGER
-,token_grants         INTEGER
-,token_issuedAt       INTEGER
+,token_expiresAt      BIGINT
+,token_grants         BIGINT
+,token_issuedAt       BIGINT
 ,token_createdBy      INTEGER
 );
